feat(bashrc): allow customization of 'master' branch w/ gm/gmm/gorm/grm aliases
For repos with non-master default branches, the branch used in these aliases can be customized by overriding the YAK_GIT_DEFAULT_BRANCH env var.

A convenience function, _github_get_default_branch, is provided, which uses the 'hub' CLI to make an API request to GitHub to retrieve a repo's default branch. This function is not called automatically, because we cannot assume the user has installed 'hub', nor do I think it reasonable to introduce the small delays from API calls without explicit buy-in from the user.